+++ modAionImpl/src/org/aion/zero/impl/sync/handler/BroadcastTxHandler.java
@@ -33,7 +33,7 @@
  * Bitcoinj team.
  */
 
-package org.aion.zero.impl.sync.callback;
+package org.aion.zero.impl.sync.handler;
 
 import java.util.ArrayList;
 import java.util.List;
