+++ modAionImpl/src/org/aion/zero/impl/sync/handler/ReqBlocksHeadersHandler.java
@@ -33,25 +33,33 @@
  * Bitcoinj team.
  */
 
-package org.aion.zero.impl.sync.callback;
+package org.aion.zero.impl.sync.handler;
 
 import java.util.List;
-
+import org.aion.zero.impl.core.IAionBlockchain;
 import org.aion.p2p.Ctrl;
 import org.aion.p2p.Handler;
 import org.aion.p2p.IP2pMgr;
 import org.aion.p2p.Ver;
-import org.aion.zero.impl.core.IAionBlockchain;
 import org.aion.zero.impl.sync.Act;
-import org.aion.zero.impl.sync.msg.ReqBlocksBodies;
-import org.aion.zero.impl.sync.msg.ResBlocksBodies;
+import org.aion.zero.impl.sync.msg.ReqBlocksHeaders;
+import org.aion.zero.impl.sync.msg.ResBlocksHeaders;
+import org.aion.mcf.types.BlockIdentifier;
+import org.aion.zero.types.A0BlockHeader;
 import org.slf4j.Logger;
 
 /**
+ *
  * @author chris
- * handler for request block bodies broadcasted from network
+ * handler for request block headers from network
+ *
  */
-public final class ReqBlocksBodiesHandler extends Handler {
+public final class ReqBlocksHeadersHandler extends Handler {
+
+    /**
+     * self guardian
+     */
+    private final static int max_headers = 2000;
 
     private final Logger log;
 
@@ -59,8 +67,8 @@ public final class ReqBlocksBodiesHandler extends Handler {
 
     private final IP2pMgr p2pMgr;
 
-    public ReqBlocksBodiesHandler(final Logger _log, final IAionBlockchain _blockchain, final IP2pMgr _p2pMgr) {
-        super(Ver.V0, Ctrl.SYNC, Act.REQ_BLOCKS_BODIES);
+    public ReqBlocksHeadersHandler(final Logger _log, final IAionBlockchain _blockchain, final IP2pMgr _p2pMgr) {
+        super(Ver.V0, Ctrl.SYNC, Act.REQ_BLOCKS_HEADERS);
         this.log = _log;
         this.blockchain = _blockchain;
         this.p2pMgr = _p2pMgr;
@@ -68,14 +76,18 @@ public final class ReqBlocksBodiesHandler extends Handler {
 
     @Override
     public void receive(int _nodeIdHashcode, String _displayId, final byte[] _msgBytes) {
-        ReqBlocksBodies reqBlocks = ReqBlocksBodies.decode(_msgBytes);
-        if (reqBlocks != null) {
-            List<byte[]> blockBodies = this.blockchain.getListOfBodiesByHashes(reqBlocks.getBlocksHashes());
-            this.p2pMgr.send(_nodeIdHashcode, new ResBlocksBodies(blockBodies));
-            this.log.debug("<req-bodies req-take={} res-take={} from-node={}>", reqBlocks.getBlocksHashes().size(),
-                    blockBodies.size(), _displayId);
-
+        ReqBlocksHeaders reqHeaders = ReqBlocksHeaders.decode(_msgBytes);
+        if (reqHeaders != null) {
+            long fromBlock = reqHeaders.getFromBlock();
+            int take = reqHeaders.getTake();
+            this.log.debug("<req-headers from-block={} take={} from-node={}>", fromBlock, take,
+                    _displayId);
+            List<A0BlockHeader> headers = this.blockchain.getListOfHeadersStartFrom(
+                    new BlockIdentifier(null, fromBlock), 0, Math.min(take, max_headers), false);
+            ResBlocksHeaders rbhs = new ResBlocksHeaders(headers);
+            this.p2pMgr.send(_nodeIdHashcode, rbhs);
         } else
-            this.log.error("<req-bodies decode-msg>");
+            this.log.error("<req-headers decode-msg msg-bytes={} from-node={}>",
+                    _msgBytes == null ? 0 : _msgBytes.length, _nodeIdHashcode);
     }
 }
