+++ modAionImpl/src/org/aion/zero/impl/sync/handler/ReqStatusHandler.java
@@ -33,62 +33,53 @@
  * Bitcoinj team.
  */
 
-package org.aion.zero.impl.sync.callback;
+package org.aion.zero.impl.sync.handler;
 
-import java.util.List;
+import org.aion.zero.impl.core.IAionBlockchain;
 import org.aion.p2p.Ctrl;
 import org.aion.p2p.Handler;
+import org.aion.p2p.IP2pMgr;
 import org.aion.p2p.Ver;
 import org.aion.zero.impl.sync.Act;
-import org.aion.zero.impl.sync.SyncMgr;
-import org.aion.zero.impl.sync.msg.ResBlocksHeaders;
-import org.aion.zero.types.A0BlockHeader;
+import org.aion.zero.impl.sync.msg.ResStatus;
 import org.slf4j.Logger;
 
 /**
- *
  * @author chris
- * handler for block headers response from network
- *
+ * handler for status request from network
  */
-public final class ResBlocksHeadersCallback extends Handler {
+public final class ReqStatusHandler extends Handler {
 
     private final Logger log;
 
-    private final SyncMgr syncMgr;
+    private IAionBlockchain chain;
 
-    public ResBlocksHeadersCallback(final Logger _log, final SyncMgr _syncMgr) {
-        super(Ver.V0, Ctrl.SYNC, Act.RES_BLOCKS_HEADERS);
-        this.syncMgr = _syncMgr;
+    private IP2pMgr mgr;
+
+    private byte[] genesisHash;
+
+    public ReqStatusHandler(final Logger _log, final IAionBlockchain _chain, final IP2pMgr _mgr, final byte[] _genesisHash) {
+        super(Ver.V0, Ctrl.SYNC, Act.REQ_STATUS);
         this.log = _log;
+        this.chain = _chain;
+        this.mgr = _mgr;
+        this.genesisHash = _genesisHash;
     }
 
     @Override
-    public void receive(int _nodeIdHashcode, String _displayId, final byte[] _msgBytes) {
-        if(_msgBytes == null || _msgBytes.length == 0)
-            return;
-        ResBlocksHeaders resHeaders = ResBlocksHeaders.decode(_msgBytes);
-        if(resHeaders != null) {
-            List<A0BlockHeader> headers = resHeaders.getHeaders();
-            if(headers != null && headers.size() > 0){
-                this.log.debug(
-                        "<res-headers from-block={} take={} from-node={}>",
-                        headers.get(0).getNumber(),
-                        headers.size(),
-                        _displayId
-                );
-                this.syncMgr.validateAndAddHeaders(_nodeIdHashcode, _displayId, headers);
-            } else {
-                this.log.error(
-                    "<res-headers empty-headers from-node={} >",
-                    _displayId
-                );
-            }
-        } else
-            this.log.error(
-                    "<res-headers decode-msg msg-bytes={} from-node={} >",
-                    _msgBytes.length,
-                    _displayId
-            );
+    public void receive(int _nodeIdHashcode, String _displayId, byte[] _msg) {
+        this.log.debug(
+                "<req-status from-node={}>",
+                _displayId
+        );
+        this.mgr.send(
+                _nodeIdHashcode,
+                new ResStatus(
+                        this.chain.getBestBlock().getNumber(),
+                        this.chain.getTotalDifficulty().toByteArray(),
+                        this.chain.getBestBlockHash(),
+                        this.genesisHash
+                )
+        );
     }
 }
